listen socket setup

+++ record.py
@@ -5,6 +5,7 @@ import ffmpy
 import datetime
 import os
 import time
+import socket
 
 ##TO DO:
 #setup cron jobs to call above jobs each hour 1 minute before the hour
@@ -18,6 +19,23 @@ filename = 'rnznews_'+date_time+'.wav'
 #is there a bug in the way protocol_whitelist is parsed? Last option always ignored!
 ffmpeg_globals = "-y -hide_banner -protocol_whitelist 'file,udp,rtp,https' -v warning"
 ffmpeg_record_string = "-c:a pcm_s24be -r:a 48000 -ac 2 -t 20:00"
+
+def listen():
+	TCP_IP = '127.0.0.1'
+	TCP_PORT = 5005
+	BUFFER_SIZE = 20  # Normally 1024, but we want fast response
+	s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+	s.bind((TCP_IP, TCP_PORT))
+	s.listen(1)
+	conn, addr = s.accept()
+	print 'Connection address:', addr
+	while 1:
+    	data = conn.recv(BUFFER_SIZE)
+    	if not data: break
+    	print "received data:", data
+    	#conn.send(data)  # echo
+	conn.close()
+	return(data)
 
 recorder = ffmpy.FFmpeg(global_options=ffmpeg_globals,inputs={sdp_file : ffmpeg_record_string},outputs={(wav_dir+filename) : None })
 
@@ -37,7 +55,11 @@ if __name__ == '__main__':
 		print "waiting ...4"
 		time.sleep(1)
 		print "waiting ...5"
-		recorder.process.terminate()
+		control = listen()
+		if control == 'terminate':
+			recorder.process.terminate()
+		else:
+			print "control:{}".format(control)
 		#print "filename:", filename
 	except KeyboardInterrupt:
 		print "manually interrupted!"
